requirements.txt
tornado
aiohttp
comfyui @ git+https://github.com/hiddenswitch/ComfyUI.git